Determining remote IP address by remote client layers #792
Introduced `@TrustedClientOnly` annotation for service beans that should
not be available from non-trusted clients.
AuthenticationServiceBean now checks if client is trusted and replaces
IP address with the address of client in credentials objects if client
is not trusted.
TrustedAccessRequiredException is thrown in case of non-trusted access
to the annotated classes and their methods.